+++ asm/code-save.s
@@ -2,268 +2,6 @@
 
 	.syntax unified
 
-	thumb_func_start SaveTraps
-SaveTraps: @ 0x08085EB4
-	push {r4, lr}
-	adds r4, r0, #0
-	movs r0, #0
-	bl GetTrap
-	movs r2, #0x80
-	lsls r2, r2, #1
-	adds r1, r4, #0
-	bl WriteAndVerifySramFast
-	pop {r4}
-	pop {r0}
-	bx r0
-	.align 2, 0
-
-	thumb_func_start LoadTraps
-LoadTraps: @ 0x08085ED0
-	push {r4, r5, lr}
-	adds r5, r0, #0
-	ldr r4, .L08085EF0 @ =ReadSramFast
-	movs r0, #0
-	bl GetTrap
-	adds r1, r0, #0
-	movs r2, #0x80
-	lsls r2, r2, #1
-	ldr r3, [r4]
-	adds r0, r5, #0
-	bl _call_via_r3
-	pop {r4, r5}
-	pop {r0}
-	bx r0
-	.align 2, 0
-.L08085EF0: .4byte ReadSramFast
-
-	thumb_func_start GetLastSuspendSlotId
-GetLastSuspendSlotId: @ 0x08085EF4
-	push {lr}
-	sub sp, #0x20
-	mov r0, sp
-	bl LoadGlobalSaveInfo
-	mov r0, sp
-	ldrb r0, [r0, #0x1f]
-	cmp r0, #1
-	beq .L08085F0A
-	movs r0, #0
-	b .L08085F0C
-.L08085F0A:
-	movs r0, #1
-.L08085F0C:
-	add sp, #0x20
-	pop {r1}
-	bx r1
-	.align 2, 0
-
-	thumb_func_start GetNextSuspendSaveId
-GetNextSuspendSaveId: @ 0x08085F14
-	push {lr}
-	bl GetLastSuspendSlotId
-	adds r1, r0, #0
-	movs r0, #1
-	subs r0, r0, r1
-	pop {r1}
-	bx r1
-
-	thumb_func_start ChangeSuspendSlotId
-ChangeSuspendSlotId: @ 0x08085F24
-	push {lr}
-	sub sp, #0x20
-	mov r0, sp
-	bl LoadGlobalSaveInfo
-	mov r1, sp
-	movs r2, #0
-	mov r0, sp
-	ldrb r0, [r0, #0x1f]
-	cmp r0, #0
-	bne .L08085F3C
-	movs r2, #1
-.L08085F3C:
-	strb r2, [r1, #0x1f]
-	mov r0, sp
-	bl SaveGlobalSaveInfoNoChecksum
-	add sp, #0x20
-	pop {r0}
-	bx r0
-	.align 2, 0
-
-	thumb_func_start GetCkSum32FromSRAM
-GetCkSum32FromSRAM: @ 0x08085F4C
-	push {r4, r5, lr}
-	adds r5, r1, #0
-	ldr r1, .L08085F6C @ =ReadSramFast
-	ldr r4, .L08085F70 @ =gBuf
-	ldr r3, [r1]
-	adds r1, r4, #0
-	adds r2, r5, #0
-	bl _call_via_r3
-	adds r0, r4, #0
-	adds r1, r5, #0
-	bl Checksum32_t
-	pop {r4, r5}
-	pop {r1}
-	bx r1
-	.align 2, 0
-.L08085F6C: .4byte ReadSramFast
-.L08085F70: .4byte gBuf
-
-	thumb_func_start CkSum32SaveBlockInfo
-CkSum32SaveBlockInfo: @ 0x08085F74
-	push {r4, r5, lr}
-	adds r4, r0, #0
-	ldrh r5, [r4, #0xa]
-	ldrh r0, [r4, #8]
-	bl SramOffsetToPointer
-	adds r1, r5, #0
-	bl GetCkSum32FromSRAM
-	ldr r1, [r4, #0xc]
-	cmp r1, r0
-	bne .L08085F90
-	movs r0, #1
-	b .L08085F92
-.L08085F90:
-	movs r0, #0
-.L08085F92:
-	pop {r4, r5}
-	pop {r1}
-	bx r1
-
-	thumb_func_start GenerateSaveBlockInfoCkSum32
-GenerateSaveBlockInfoCkSum32: @ 0x08085F98
-	push {r4, r5, lr}
-	adds r4, r0, #0
-	ldrh r5, [r4, #0xa]
-	ldrh r0, [r4, #8]
-	bl SramOffsetToPointer
-	adds r1, r5, #0
-	bl GetCkSum32FromSRAM
-	str r0, [r4, #0xc]
-	pop {r4, r5}
-	pop {r0}
-	bx r0
-	.align 2, 0
-
-	thumb_func_start ComputeSaveDataCkSum32
-ComputeSaveDataCkSum32: @ 0x08085FB4
-	push {r4, r5, r6, lr}
-	movs r6, #0
-	ldr r4, .L0808607C @ =gUnitArrayBlue
-	movs r5, #0x33
-.L08085FBC:
-	ldr r0, [r4]
-	cmp r0, #0
-	beq .L08085FD4
-	movs r0, #0
-	str r0, [r4, #0x3c]
-	adds r0, r4, #0
-	movs r1, #0x24
-	bl GetCkSum32FromSRAM
-	adds r0, r6, r0
-	lsls r0, r0, #0x10
-	lsrs r6, r0, #0x10
-.L08085FD4:
-	adds r4, #0x48
-	subs r5, #1
-	cmp r5, #0
-	bge .L08085FBC
-	ldr r4, .L08086080 @ =gUnitArrayRed
-	movs r5, #0x31
-.L08085FE0:
-	ldr r0, [r4]
-	cmp r0, #0
-	beq .L08085FF8
-	movs r0, #0
-	str r0, [r4, #0x3c]
-	adds r0, r4, #0
-	movs r1, #0x24
-	bl GetCkSum32FromSRAM
-	adds r0, r6, r0
-	lsls r0, r0, #0x10
-	lsrs r6, r0, #0x10
-.L08085FF8:
-	adds r4, #0x48
-	subs r5, #1
-	cmp r5, #0
-	bge .L08085FE0
-	ldr r4, .L08086084 @ =gUnitArrayGreen
-	movs r5, #9
-.L08086004:
-	ldr r0, [r4]
-	cmp r0, #0
-	beq .L0808601C
-	movs r0, #0
-	str r0, [r4, #0x3c]
-	adds r0, r4, #0
-	movs r1, #0x24
-	bl GetCkSum32FromSRAM
-	adds r0, r6, r0
-	lsls r0, r0, #0x10
-	lsrs r6, r0, #0x10
-.L0808601C:
-	adds r4, #0x48
-	subs r5, #1
-	cmp r5, #0
-	bge .L08086004
-	bl GetPermanentFlagBits
-	adds r4, r0, #0
-	bl GetPermanentFlagBitsSize
-	adds r1, r0, #0
-	lsrs r0, r1, #0x1f
-	adds r1, r1, r0
-	asrs r1, r1, #1
-	adds r0, r4, #0
-	bl GetCkSum32FromSRAM
-	adds r0, r6, r0
-	lsls r0, r0, #0x10
-	lsrs r6, r0, #0x10
-	bl GetChapterFlagBits
-	adds r4, r0, #0
-	bl GetChapterFlagBitsSize
-	adds r1, r0, #0
-	lsrs r0, r1, #0x1f
-	adds r1, r1, r0
-	asrs r1, r1, #1
-	adds r0, r4, #0
-	bl GetCkSum32FromSRAM
-	adds r0, r6, r0
-	lsls r0, r0, #0x10
-	lsrs r6, r0, #0x10
-	movs r0, #0
-	bl GetTrap
-	movs r1, #0x80
-	bl GetCkSum32FromSRAM
-	adds r0, r6, r0
-	lsls r0, r0, #0x10
-	lsrs r6, r0, #0x10
-	adds r0, r6, #0
-	pop {r4, r5, r6}
-	pop {r1}
-	bx r1
-	.align 2, 0
-.L0808607C: .4byte gUnitArrayBlue
-.L08086080: .4byte gUnitArrayRed
-.L08086084: .4byte gUnitArrayGreen
-
-	thumb_func_start func_fe6_08086088
-func_fe6_08086088: @ 0x08086088
-	sub sp, #8
-	add sp, #8
-	bx lr
-	.align 2, 0
-
-	thumb_func_start VerifySaveBlockInfoByIndex2
-VerifySaveBlockInfoByIndex2: @ 0x08086090
-	push {lr}
-	adds r1, r0, #0
-	movs r0, #0
-	bl LoadSaveBlockInfo
-	lsls r0, r0, #0x18
-	asrs r0, r0, #0x18
-	pop {r1}
-	bx r1
-	.align 2, 0
-
 	thumb_func_start func_fe6_080860A4
 func_fe6_080860A4: @ 0x080860A4
 	push {r4, r5, r6, r7, lr}
