+++ src/main.c
@@ -9,6 +9,7 @@
 #include "gamecontroller.h"
 #include "bm.h"
 #include "mu.h"
+#include "save.h"
 
 #include "m4a.h"
